Add routing
Uses svelte-spa-routing with hash based routing so no problems with github pages
