+++ package.json
@@ -9,6 +9,9 @@
     "preview": "vite preview",
     "lint": "eslint \"src/**/*.{svelte,js}\""
   },
+  "dependencies": {
+    "svelte-spa-router": "^3.2.0"
+  },
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^1.0.0-next.39",
     "autoprefixer": "^10.4.2",
