+++ src/lib/Logo.svelte
@@ -2,4 +2,6 @@
   import logo from '../assets/tcflix_logo.png';
 </script>
 
-<img src={logo} alt="logo" width="119" height="45">
+<a href="#/">
+    <img src={logo} alt="logo" width="119" height="45">
+</a>
