+++ src/pages/About.svelte
@@ -1,0 +1,3 @@
+<p>Hello there</p>
+<img src="https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fi.redd.it%2Fcaa6g5u9amoy.jpg&f=1&nofb=1"
+     alt="obiwan"/>
